requirements.txt
blinker==1.6.2
certifi==2023.5.7
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.3.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.3.3
python-dotenv==1.0.0
Werkzeug==2.3.4
zipp==3.15.0
